Move credential for mDL out to separate library
