+++ vclib-openid/src/commonTest/kotlin/KotestConfig.kt
@@ -6,5 +6,7 @@ object KotestConfig : AbstractProjectConfig() {
     init {
         Napier.takeLogarithm()
         Napier.base(DebugAntilog())
+        at.asitplus.wallet.mdl.Initializer.initWithVcLib()
+        at.asitplus.wallet.eupid.Initializer.initWithVcLib()
     }
 }
