+++ vclib-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/oidc/OidcSiopWalletScopeSupportTest.kt
@@ -15,8 +15,9 @@ import at.asitplus.wallet.lib.data.dif.ConstraintField
 import at.asitplus.wallet.lib.data.dif.InputDescriptor
 import at.asitplus.wallet.lib.data.dif.PresentationDefinition
 import at.asitplus.wallet.lib.data.dif.SchemaReference
-import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements
+import at.asitplus.wallet.mdl.MobileDrivingLicenceDataElements
 import at.asitplus.wallet.lib.oidvci.OAuth2Exception
+import at.asitplus.wallet.mdl.MobileDrivingLicenceScheme
 import com.benasher44.uuid.uuid4
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.FreeSpec
@@ -48,13 +49,13 @@ class OidcSiopWalletScopeSupportTest : FreeSpec({
                 id = uuid4().toString(),
                 inputDescriptors = listOf(
                     InputDescriptor(
-                        id = ConstantIndex.MobileDrivingLicence2023.isoDocType,
+                        id = MobileDrivingLicenceScheme.isoDocType,
                         constraints = Constraint(
                             fields = listOf(
                                 ConstraintField(
                                     path = listOf(
                                         NormalizedJsonPath(
-                                            NormalizedJsonPathSegment.NameSegment(ConstantIndex.MobileDrivingLicence2023.isoNamespace),
+                                            NormalizedJsonPathSegment.NameSegment(MobileDrivingLicenceScheme.isoNamespace),
                                             NormalizedJsonPathSegment.NameSegment(MobileDrivingLicenceDataElements.GIVEN_NAME),
                                         ).toString()
                                     ),
@@ -62,7 +63,7 @@ class OidcSiopWalletScopeSupportTest : FreeSpec({
                             )
                         ),
                         schema = listOf(
-                            SchemaReference(ConstantIndex.MobileDrivingLicence2023.schemaUri)
+                            SchemaReference(MobileDrivingLicenceScheme.schemaUri)
                         )
                     )
                 )
@@ -154,7 +155,7 @@ class OidcSiopWalletScopeSupportTest : FreeSpec({
                 holderAgent.storeCredentials(
                     issuerAgent.issueCredential(
                         subjectPublicKey = holderCryptoService.publicKey,
-                        attributeTypes = listOf(ConstantIndex.MobileDrivingLicence2023.vcType),
+                        attributeTypes = listOf(MobileDrivingLicenceScheme.isoNamespace),
                         representation = ConstantIndex.CredentialRepresentation.ISO_MDOC
                     ).toStoreCredentialInput()
                 )
