+++ vclib/src/commonMain/kotlin/at/asitplus/wallet/lib/data/AriesGoalCodeParser.kt
@@ -5,7 +5,6 @@ object AriesGoalCodeParser {
 
     init {
         registerGoalCode(ConstantIndex.AtomicAttribute2023)
-        registerGoalCode(ConstantIndex.MobileDrivingLicence2023)
     }
 
     fun parseGoalCode(goalCode: String) = when (goalCode) {
