+++ vclib/src/commonMain/kotlin/at/asitplus/wallet/lib/data/AttributeIndex.kt
@@ -6,7 +6,6 @@ object AttributeIndex {
 
     init {
         schemeSet += ConstantIndex.AtomicAttribute2023
-        schemeSet += ConstantIndex.MobileDrivingLicence2023
     }
 
     internal fun registerAttributeType(scheme: ConstantIndex.CredentialScheme) {
