+++ vclib/src/commonMain/kotlin/at/asitplus/wallet/lib/data/Json.kt
@@ -2,10 +2,8 @@ package at.asitplus.wallet.lib.data
 
 import at.asitplus.wallet.lib.JsonValueEncoder
 import at.asitplus.wallet.lib.data.Json.serializersModules
-import at.asitplus.wallet.lib.iso.DrivingPrivilege
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonElement
-import kotlinx.serialization.json.encodeToJsonElement
 import kotlinx.serialization.modules.SerializersModule
 import kotlinx.serialization.modules.polymorphic
 import kotlinx.serialization.modules.subclass
@@ -17,12 +15,6 @@ object Json {
 
     fun registerSerializersModule(scheme: ConstantIndex.CredentialScheme, module: SerializersModule) {
         serializersModules[scheme] = module
-    }
-
-    init {
-        register {
-            if (it is DrivingPrivilege) jsonSerializer.encodeToJsonElement(it) else null
-        }
     }
 
     fun register(function: JsonValueEncoder) {
